CS-4 // order items

+++ src/main/java/com/javalab/clothshop/service/product/ProductRemovalServiceImpl.java
@@ -1,6 +1,7 @@
 package com.javalab.clothshop.service.product;
 
 import com.javalab.clothshop.repository.ProductRepository;
+import com.javalab.clothshop.repository.exception.ProductNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +13,10 @@ public class ProductRemovalServiceImpl implements ProductRemovalService {
 
     @Override
     public void removeById(Long id) {
-        productRepository.deleteById(id);
+        if (productRepository.existsById(id)) {
+            productRepository.deleteById(id);
+        } else {
+            throw new ProductNotFoundException("Could not find product with id: " + id);
+        }
     }
 }
